+++ src/main/java/co/aisaac/nes_java/apu/NESConstants.java
@@ -5,20 +5,20 @@ import static co.aisaac.nes_java.cpu.CPU.CPUFrequency;
 public class NESConstants {
     public static final double frameCounterRate = CPUFrequency / 240.0;
 
-    public static final byte[] lengthTable = new byte[]{
+    public static final int /*byte*/[] lengthTable = new int /*byte*/[]{
             // todo
-            10, (byte) 254, 20, 2, 40, 4, 80, 6, (byte) 160, 8, 60, 10, 14, 12, 26, 14,
-            12, 16, 24, 18, 48, 20, 96, 22, (byte) 192, 24, 72, 26, 16, 28, 32, 30,
+            10, 254, 20, 2, 40, 4, 80, 6, 160, 8, 60, 10, 14, 12, 26, 14,
+            12, 16, 24, 18, 48, 20, 96, 22, 192, 24, 72, 26, 16, 28, 32, 30,
     };
 
-    public static final byte[][] dutyTable = new byte[][]{
+    public static final int /*byte*/[][] dutyTable = new int /*byte*/[][]{
             {0, 1, 0, 0, 0, 0, 0, 0},
             {0, 1, 1, 0, 0, 0, 0, 0},
             {0, 1, 1, 1, 1, 0, 0, 0},
             {1, 0, 0, 1, 1, 1, 1, 1},
     };
 
-    public static final byte[] triangleTable = new byte[]{
+    public static final int /*byte*/[] triangleTable = new int /*byte*/[]{
             15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1, 0,
             0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15,
     };
@@ -27,8 +27,8 @@ public class NESConstants {
             4, 8, 16, 32, 64, 96, 128, 160, 202, 254, 380, 508, 762, 1016, 2034, 4068,
     };
 
-    public static final byte[] dmcTable = new byte[]{
-            (byte) 214, (byte) 190, (byte) 170, (byte) 160, (byte) 143, (byte) 127, (byte) 113, (byte) 107, (byte) 95, (byte) 80, (byte) 71, (byte) 64, (byte) 53, (byte) 42, (byte) 36, (byte) 27,
+    public static final int /*byte*/[] dmcTable = new int /*byte*/[]{
+            (int /*byte*/) 214, (int /*byte*/) 190, (int /*byte*/) 170, (int /*byte*/) 160, (int /*byte*/) 143, (int /*byte*/) 127, (int /*byte*/) 113, (int /*byte*/) 107, (int /*byte*/) 95, (int /*byte*/) 80, (int /*byte*/) 71, (int /*byte*/) 64, (int /*byte*/) 53, (int /*byte*/) 42, (int /*byte*/) 36, (int /*byte*/) 27,
     };
 
     public static final float[] pulseTable = new float[31];
